% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_protein.R
\name{dna_protein}
\alias{dna_protein}
\title{This is the function for dna to protein}
\usage{
dna_protein(dna)
}
\arguments{
\item{dna}{vector of dna chr}
}
\description{
This is the function for dna to protein
}
\author{
liripo
}
