% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_protein.R
\name{dna_to_protein}
\alias{dna_to_protein}
\title{This is the function for dna to protein}
\usage{
dna_to_protein(dna)
}
\description{
This is the function for dna to protein
}
\examples{
dna_to_protein(c("ATG","GCC"))
}
\author{
liripo
}
