% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{rmb_pal}
\alias{rmb_pal}
\alias{scale_color_rmb}
\alias{scale_colour_rmb}
\alias{scale_fill_rmb}
\title{rmb pal}
\usage{
rmb_pal(money = "100", alpha = 1, reverse = TRUE)

scale_color_rmb(..., discrete = TRUE, money = "100", alpha = 1, reverse = TRUE)

scale_colour_rmb(
  ...,
  discrete = TRUE,
  money = "100",
  alpha = 1,
  reverse = TRUE
)

scale_fill_rmb(..., discrete = TRUE, money = "100", alpha = 1, reverse = TRUE)
}
\arguments{
\item{money}{rmb money character,It's can be set "100","50",..."1hairs"}

\item{alpha}{alpha}

\item{reverse}{logical}

\item{...}{see scale_color_gradientn}

\item{discrete}{logical;whether discrete,the deault is T.}
}
\description{
rmb color
}
\author{
liripo
}
